Split the core functionality out of SimpleUI
