+++ Sprocket/Sprocket.h
@@ -14,8 +14,9 @@
 #include "UI/Font/Font.h"
 #include "UI/Font/Glyph.h"
 
-#include "UI/DevUI.h"
+#include "UI/UIEngine.h"
 #include "UI/SimpleUI.h"
+#include "UI/DevUI.h"
 
 // EVENTS
 #include "Events/Event.h"
