+++ Sprocket/UI/SimpleUI.cpp
@@ -77,145 +77,32 @@ template <typename T> T Interpolate(
 
 SimpleUI::SimpleUI(Window* window)
     : d_window(window)
-    , d_shader("Resources/Shaders/SimpleUI.vert",
-               "Resources/Shaders/SimpleUI.frag")
-    , d_font("Resources/Fonts/Coolvetica.ttf")
+    , d_engine(window)
 {
     d_keyboard.ConsumeAll(false);
-
-    BufferLayout layout(sizeof(BufferVertex));
-    layout.AddAttribute(DataType::FLOAT, 2);
-    layout.AddAttribute(DataType::FLOAT, 4);
-    layout.AddAttribute(DataType::FLOAT, 2);
-    d_buffer.SetBufferLayout(layout);
-}
-
-Maths::vec4 SimpleUI::ApplyOffset(const Maths::vec4& region)
-{
-    if (d_currentPanel) {
-        Maths::vec4 quad = region;
-        quad.x += d_currentPanel->region.x;
-        quad.y += d_currentPanel->region.y;
-        return quad;
-    }
-    return region;
-}
-
-std::string SimpleUI::MangleName(const std::string& name)
-{
-    return d_currentPanel->name + "##" + name;
 }
 
 void SimpleUI::OnEvent(Event& event)
 {
     d_keyboard.OnEvent(event);
     d_mouse.OnEvent(event);
+    d_engine.OnEvent(event);
 }
 
 void SimpleUI::OnUpdate(double dt)
 {
     d_mouse.OnUpdate();
-    d_time += dt;
-    d_clickedTime += dt;
-    d_hoveredTime += dt;
-
-    if (d_mouse.IsButtonReleased(Mouse::LEFT)) {
-        d_clickedTime = 0.0;
-        if (d_clicked > 0) {
-            d_widgetTimes[d_clicked].unclickedTime = d_time;
-            d_clicked = 0;
-        }
-    }
+    d_engine.OnUpdate(dt);
 }
 
 void SimpleUI::StartFrame()
 {
-    d_panels.clear();
-    d_currentPanel = nullptr;
+    d_engine.StartFrame();
 }
 
 void SimpleUI::EndFrame()
 {
-    assert(!d_currentPanel);
-
-    bool foundHovered = false;
-    bool foundClicked = false;
-
-    std::size_t moveToFront = 0;
-
-    for (const auto& panelHash : Reversed(d_panelOrder)) {
-        const auto& panel = d_panels[panelHash];
-
-        for (const auto& quad : Reversed(panel.widgetRegions)) {
-            std::size_t hash = quad.hash;
-            auto hovered = d_mouse.InRegion(quad.region.x, quad.region.y, quad.region.z, quad.region.w);
-            auto clicked = hovered && d_mouse.IsButtonClicked(Mouse::LEFT);
-
-            if (!foundClicked && ((d_clicked == hash) || clicked)) {
-                foundClicked = true;
-                moveToFront = panelHash;
-                if (d_clicked != hash) {
-                    d_widgetTimes[d_clicked].unclickedTime = d_time;
-                    d_widgetTimes[hash].clickedTime = d_time;
-                    d_clicked = hash;
-                    d_onClick = hash;
-                    d_clickedTime = 0.0;
-                }
-            }
-            
-            if (!foundHovered && hovered) {
-                foundHovered = true;
-                if (d_hovered != hash) {
-                    d_widgetTimes[d_hovered].unhoveredTime = d_time;
-                    d_widgetTimes[hash].hoveredTime = d_time;
-                    d_hovered = hash;
-                    d_onHover = hash;
-                    d_hoveredTime = 0.0;
-                }
-            }
-        }
-    }
-
-    if (foundHovered == false) {
-        d_hoveredTime = 0.0;
-        if (d_hovered > 0) {
-            d_widgetTimes[d_hovered].unhoveredTime = d_time;
-            d_hovered = 0;
-        }
-    }
-
-    Sprocket::RenderContext rc;
-    rc.AlphaBlending(true);
-    rc.FaceCulling(false);
-    rc.DepthTesting(false);
-
-    float w = (float)d_window->Width();
-    float h = (float)d_window->Height();
-    auto proj = Maths::Ortho(0, w, h, 0);
-    d_shader.Bind();
-    d_shader.LoadUniform("u_proj_matrix", proj);
-
-    d_buffer.Bind();
-    for (const auto& panelHash : d_panelOrder) {
-        const auto& panel = d_panels[panelHash];
-        d_shader.LoadUniformInt("texture_channels", 1);
-        Texture::White().Bind();
-        d_buffer.Draw(panel.quadVertices, panel.quadIndices);
-        d_font.Bind();
-        d_buffer.Draw(panel.textVertices, panel.textIndices);
-        for (const auto& cmd : panel.extraCommands) {
-            d_shader.LoadUniformInt("texture_channels", cmd.texture->GetChannels());
-            cmd.texture->Bind();
-            d_buffer.Draw(cmd.vertices, cmd.indices);
-        }
-    }
-    d_buffer.Unbind();
-
-    if (moveToFront > 0) {
-        auto toMove = std::find(d_panelOrder.begin(), d_panelOrder.end(), moveToFront);
-        d_panelOrder.erase(toMove);
-        d_panelOrder.push_back(moveToFront);
-    }
+    d_engine.EndFrame();
 }
 
 bool SimpleUI::StartPanel(
@@ -224,32 +111,9 @@ bool SimpleUI::StartPanel(
     bool* active,
     bool* draggable)
 {
-    assert(!d_currentPanel);
-    assert(region != nullptr);
-    assert(active != nullptr);
+    d_engine.StartPanel(name, region, active, draggable);
 
-    if (*active) {
-        std::size_t hash = std::hash<std::string>{}(name);
-
-        auto it = std::find(d_panelOrder.begin(), d_panelOrder.end(), hash);
-        if (it == d_panelOrder.end()) {
-            d_panelOrder.push_back(hash);
-        }
-        
-        auto& panel = d_panels[hash];
-        panel.name = name;
-        panel.hash = hash;
-        panel.region = *region;
-
-        d_currentPanel = &panel;
-        
-        auto info = RegisterWidget(name, *region);
-
-        if (info.mouseDown && *draggable) {
-            region->x += d_mouse.GetMouseOffset().x;
-            region->y += d_mouse.GetMouseOffset().y;
-        }
-
+    if(*active) {
         float thickness = 5.0f;
         auto border = *region;
         border.x -= thickness;
@@ -257,8 +121,8 @@ bool SimpleUI::StartPanel(
         border.z += 2.0f * thickness;
         border.w += 2.0f * thickness;
 
-        DrawQuad(d_theme.backgroundColour * 0.35f, border);
-        DrawQuad(d_theme.backgroundColour * 0.7f, *region);
+        d_engine.DrawQuad(d_theme.backgroundColour * 0.35f, border);
+        d_engine.DrawQuad(d_theme.backgroundColour * 0.7f, *region);
     }
 
     return *active;
@@ -266,102 +130,13 @@ bool SimpleUI::StartPanel(
 
 void SimpleUI::EndPanel()
 {
-    assert(d_currentPanel);
-    d_currentPanel = nullptr;
-}
-
-void SimpleUI::DrawQuad(const Maths::vec4& colour,
-                        const Maths::vec4& region)
-{
-    assert(d_currentPanel);
-    auto copy = region;
-    float x = copy.x;
-    float y = copy.y;
-    float width = copy.z;
-    float height = copy.w;
-
-    auto& cmd = d_panels[d_currentPanel->hash];
-
-    auto col = colour;
-    col.a = 1;
-
-    std::size_t index = cmd.quadVertices.size();
-    cmd.quadVertices.push_back({{x,         y},          col});
-    cmd.quadVertices.push_back({{x + width, y},          col});
-    cmd.quadVertices.push_back({{x,         y + height}, col});
-    cmd.quadVertices.push_back({{x + width, y + height}, col});
-
-    cmd.quadIndices.push_back(index + 0);
-    cmd.quadIndices.push_back(index + 1);
-    cmd.quadIndices.push_back(index + 2);
-    cmd.quadIndices.push_back(index + 2);
-    cmd.quadIndices.push_back(index + 1);
-    cmd.quadIndices.push_back(index + 3);
-}
-
-void SimpleUI::DrawText(
-    const std::string& text,
-    float size,
-    const Maths::vec4& region)
-{
-    assert(d_currentPanel);
-
-    auto copy = region;
-    Maths::vec4 colour = {1.0, 1.0, 1.0, 1.0};
-
-    Glyph first = d_font.GetGlyph(text.front(), size);
-    auto textInfo = GetTextInfo(d_font, size, text);
-
-    Maths::vec2 pen = {
-        region.x + (copy.z - textInfo.width) / 2.0f,
-        region.y + (copy.w - first.height) / 2.0f
-    };
-
-    pen.x -= first.offset.x;
-    pen.y += first.offset.y;
-    
-    for (std::size_t i = 0; i != text.size(); ++i) {
-        auto glyph = d_font.GetGlyph(text[i], size);
-
-        if (i > 0) {
-            pen.x += d_font.GetKerning(text[i-1], text[i], size);
-        }
-
-        float xPos = pen.x + glyph.offset.x;
-        float yPos = pen.y - glyph.offset.y;
-
-        float width = glyph.width;
-        float height = glyph.height;
-
-        float x = glyph.texture.x;
-        float y = glyph.texture.y;
-        float w = glyph.texture.z;
-        float h = glyph.texture.w;
-
-        pen += glyph.advance;
-
-        auto& cmd = d_panels[d_currentPanel->hash];
-
-        unsigned int index = cmd.textVertices.size();
-        cmd.textVertices.push_back({{xPos,         yPos},          colour, {x,     y    }});
-        cmd.textVertices.push_back({{xPos + width, yPos},          colour, {x + w, y    }});
-        cmd.textVertices.push_back({{xPos,         yPos + height}, colour, {x,     y + h}});
-        cmd.textVertices.push_back({{xPos + width, yPos + height}, colour, {x + w, y + h}});
-
-        cmd.textIndices.push_back(index + 0);
-        cmd.textIndices.push_back(index + 1);
-        cmd.textIndices.push_back(index + 2);
-        cmd.textIndices.push_back(index + 2);
-        cmd.textIndices.push_back(index + 1);
-        cmd.textIndices.push_back(index + 3);
-    }
+    d_engine.EndPanel();
 }
 
 void SimpleUI::Quad(const Maths::vec4& colour, const Maths::vec4& quad)
 {
-    assert(d_currentPanel);
-    auto copy = ApplyOffset(quad);
-    DrawQuad(colour, copy);
+    auto copy = d_engine.ApplyOffset(quad);
+    d_engine.DrawQuad(colour, copy);
 }
 
 void SimpleUI::Text(
@@ -369,17 +144,15 @@ void SimpleUI::Text(
     float size,
     const Maths::vec4& quad)
 {
-    assert(d_currentPanel);
-    auto copy = ApplyOffset(quad);
-    DrawText(text, size, copy);
+    auto copy = d_engine.ApplyOffset(quad);
+    d_engine.DrawText(text, size, copy);
 }
 
 bool SimpleUI::Button(const std::string& name,
                       const Maths::vec4& region)
 {
-    assert(d_currentPanel);
-    Maths::vec4 copy = ApplyOffset(region);
-    auto info = RegisterWidget(MangleName(name), copy);
+    Maths::vec4 copy = d_engine.ApplyOffset(region);
+    auto info = d_engine.RegisterWidget(d_engine.MangleName(name), copy);
 
     Maths::vec4 hoveredRegion = copy;
     hoveredRegion.x -= 10.0f;
@@ -392,8 +165,8 @@ bool SimpleUI::Button(const std::string& name,
     Maths::vec4 colour = Interpolate(info, d_theme.baseColour, d_theme.hoveredColour, d_theme.clickedColour);
     Maths::vec4 shape = Interpolate(info, copy, hoveredRegion, clickedRegion);
     
-    DrawQuad(colour, shape);
-    DrawText(name, 36.0f, copy);
+    d_engine.DrawQuad(colour, shape);
+    d_engine.DrawText(name, 36.0f, copy);
 
     return info.onClick;
 }
@@ -402,9 +175,8 @@ bool SimpleUI::Checkbox(const std::string& name,
                         const Maths::vec4& region,
                         bool* value)
 {
-    assert(d_currentPanel);
-    Maths::vec4 copy = ApplyOffset(region);
-    auto info = RegisterWidget(MangleName(name), copy);
+    Maths::vec4 copy = d_engine.ApplyOffset(region);
+    auto info = d_engine.RegisterWidget(d_engine.MangleName(name), copy);
 
     auto unselected = Interpolate(info, d_theme.backgroundColour, d_theme.backgroundColour*1.1f, d_theme.clickedColour);
     auto selected = Interpolate(info, d_theme.baseColour, d_theme.hoveredColour, d_theme.clickedColour);;
@@ -422,8 +194,8 @@ bool SimpleUI::Checkbox(const std::string& name,
         *value = !(*value);
     }
 
-    DrawQuad(colour, copy);
-    DrawText(name, 36.0f, copy);
+    d_engine.DrawQuad(colour, copy);
+    d_engine.DrawText(name, 36.0f, copy);
 
     return *value; 
 }
@@ -432,9 +204,8 @@ void SimpleUI::Slider(const std::string& name,
                       const Maths::vec4& region,
                       float* value, float min, float max)
 {
-    assert(d_currentPanel);
-    Maths::vec4 copy = ApplyOffset(region);
-    auto info = RegisterWidget(MangleName(name), copy);
+    Maths::vec4 copy = d_engine.ApplyOffset(region);
+    auto info = d_engine.RegisterWidget(d_engine.MangleName(name), copy);
 
     float x = copy.x;
     float y = copy.y;
@@ -445,9 +216,9 @@ void SimpleUI::Slider(const std::string& name,
     Maths::vec4 rightColour = d_theme.backgroundColour;
     
     float ratio = (*value - min) / (max - min);
-    DrawQuad(leftColour, {x, y, ratio * width, height});
-    DrawQuad(rightColour, {x + ratio * width, y, (1 - ratio) * width, height});
-    DrawText(name + ": " + Maths::ToString(*value, 0), 36.0f, copy);
+    d_engine.DrawQuad(leftColour, {x, y, ratio * width, height});
+    d_engine.DrawQuad(rightColour, {x + ratio * width, y, (1 - ratio) * width, height});
+    d_engine.DrawText(name + ": " + Maths::ToString(*value, 0), 36.0f, copy);
 
     if (info.mouseDown) {
         auto mouse = d_mouse.GetMousePos();
@@ -461,14 +232,13 @@ void SimpleUI::Dragger(const std::string& name,
                        const Maths::vec4& region,
                        float* value, float speed)
 {
-    assert(d_currentPanel);
-    Maths::vec4 copy = ApplyOffset(region);
-    auto info = RegisterWidget(MangleName(name), copy);
+    Maths::vec4 copy = d_engine.ApplyOffset(region);
+    auto info = d_engine.RegisterWidget(d_engine.MangleName(name), copy);
 
     Maths::vec4 colour = Interpolate(info, d_theme.baseColour, d_theme.hoveredColour, d_theme.clickedColour);
     
-    DrawQuad(colour, copy);
-    DrawText(name + ": " + Maths::ToString(*value, 0), 36.0f, copy);
+    d_engine.DrawQuad(colour, copy);
+    d_engine.DrawText(name + ": " + Maths::ToString(*value, 0), 36.0f, copy);
 
     if (info.mouseDown) {
         *value += d_mouse.GetMouseOffset().x * speed;
@@ -479,11 +249,10 @@ void SimpleUI::Image(const std::string& name,
                      const Texture& image,
                      const Maths::vec2& position)
 {
-    assert(d_currentPanel);
     Maths::vec4 region{position.x, position.y, image.Width(), image.Height()};
-    Maths::vec4 copy = ApplyOffset(region);
+    Maths::vec4 copy = d_engine.ApplyOffset(region);
 
-    auto& cmd = d_currentPanel->extraCommands.emplace_back(DrawCommand());
+    DrawCommand cmd;
     cmd.vertices = {
         {{copy.x,          copy.y         }, {1.0, 1.0, 1.0, 1.0}, {0.0, 0.0}},
         {{copy.x + copy.z, copy.y         }, {1.0, 1.0, 1.0, 1.0}, {1.0, 0.0}},
@@ -492,42 +261,7 @@ void SimpleUI::Image(const std::string& name,
     };
     cmd.indices = {0, 1, 2, 2, 1, 3};
     cmd.texture = &image;
+    d_engine.SubmitDrawCommand(cmd);
 }
-
-WidgetInfo SimpleUI::RegisterWidget(const std::string& name,
-                                    const Maths::vec4& region)
-{
-    assert(d_currentPanel);
-    WidgetInfo info;
-    std::size_t hash = std::hash<std::string>{}(name);
-    d_panels[d_currentPanel->hash].widgetRegions.push_back({hash, region});
-
-    if (hash == d_clicked) {
-        info.mouseDown = d_clickedTime;
-    }
-
-    info.sinceUnlicked = d_time - d_widgetTimes[hash].unclickedTime;
-    info.sinceClicked = d_time - d_widgetTimes[hash].clickedTime;
-
-    if (hash == d_hovered) {
-        info.mouseOver = d_hoveredTime;
-    }
-    
-    info.sinceUnhovered = d_time - d_widgetTimes[hash].unhoveredTime;
-    info.sinceHovered = d_time - d_widgetTimes[hash].hoveredTime;
-
-    if (d_onClick == hash) { // Consume the onCLick
-        d_onClick = 0;
-        info.onClick = true;
-    }
-
-    if (d_onHover == hash) { // Consume the onHover
-        d_onHover = 0;
-        info.onHover = true;
-    }
-
-    return info;
-}
-
 
 }
